requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.10
Django==3.2.11
djangorestframework==3.13.1
idna==3.3
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
typing-extensions==4.0.1
urllib3==1.26.8
